Refactor executeCommand to remove import cycle
