+++ actions/views/create_comment.jsx
@@ -18,7 +18,7 @@ import {isPostPendingOrFailed} from 'mattermost-redux/utils/post_utils';
 
 import * as PostActions from 'actions/post_actions.jsx';
 import * as GlobalActions from 'actions/global_actions.jsx';
-import * as ChannelActions from 'actions/channel_actions.jsx';
+import {executeCommand} from 'actions/command';
 import {setGlobalItem, actionOnGlobalItemsWithPrefix} from 'actions/storage';
 import {EmojiMap} from 'stores/emoji_store.jsx';
 import {getPostDraft} from 'selectors/rhs';
@@ -97,7 +97,7 @@ export function submitReaction(postId, action, emojiName) {
 }
 
 export function submitCommand(channelId, rootId, draft) {
-    return (dispatch, getState) => {
+    return async (dispatch, getState) => {
         const state = getState();
 
         const teamId = getCurrentTeamId(state);
@@ -111,15 +111,15 @@ export function submitCommand(channelId, rootId, draft) {
 
         const {message} = draft;
 
-        return new Promise((resolve, reject) => {
-            ChannelActions.executeCommand(message, args, resolve, (err) => {
-                if (err.sendMessage) {
-                    dispatch(submitPost(channelId, rootId, draft));
-                } else {
-                    reject(err);
-                }
-            });
-        });
+        const {error} = await dispatch(executeCommand(message, args));
+
+        if (error) {
+            if (error.sendMessage) {
+                dispatch(submitPost(channelId, rootId, draft));
+            } else {
+                throw (error);
+            }
+        }
     };
 }
