+++ tests/components/dot_menu/dot_menu_empty.test.jsx
@@ -9,6 +9,7 @@ import DotMenu from 'components/dot_menu/dot_menu.jsx';
 jest.mock('utils/utils', () => {
     return {
         isMobile: jest.fn(() => false),
+        localizeMessage: jest.fn(),
     };
 });
